refactor(timeline): move timeline list into timeline module
